+++ bookalo/views.py
@@ -18,16 +18,9 @@ from decimal import Decimal
 from .funciones_producto import *
 from .funciones_chat import *
 from .funciones_report import *
+from funciones_usuario import *
 
-#Comprueba que el usuario este logeado en el sistema
-def check_user_logged_in(token):
-	try:
-		user_info = auth.get_account_info(token)
-		user_uid = user_info['users'][0]['localId']
-		user = Usuario.objects.get(uid=user_uid).update(ultima_conexion=datetime.now)
-		return True
-	except:
-		return False
+
 
 def update_last_connection(user):
 	try:
@@ -52,46 +45,20 @@ def get_user(token):
 def index(request):
 	return render(request, 'bookalo/index.html')
 
+
+
 @api_view(['POST'])
 @permission_classes((permissions.AllowAny,))
 def Login(request, format=None):
 	token = request.POST.get('token', 'nothing')
-	latitud_registro = 0.0
-	longitud_registro = 0.0
-	g = GeoIP2()
-	#ip = request.META.get('REMOTE_ADDR', None)
-	#if ip:
-	#	latitud_registro = g.city(ip)['latitude']
-	#	longitud_registro = g.city(ip)['longitude']
-	latitud_registro = 41.683490
-	longitud_registro = -0.888479
 	if request.method != 'POST':
 		return Response(status=status.HTTP_400_BAD_REQUEST)
-	if token == 'nothing':
-		return Response(status=status.HTTP_400_BAD_REQUEST)
-	else:
-		try:
-			user_info = auth.get_account_info(token)
-			user_uid = user_info['users'][0]['localId']
-			name = user_info['users'][0]['email'].split("@")[0]
-		except:
+	else
+		retorno = usuario_login(token)
+		if retorno = 'Error'
 			return Response(status=status.HTTP_404_NOT_FOUND)
-		
-		try:
-			user = Usuario.objects.get(uid=user_uid)
-			user.ultima_conexion = datetime.now()
-			user.save(update_fields=['ultima_conexion'])
-
-			if user.esta_baneado:
-				return Response(UserSerializer(user).data, status=status.HTTP_401_UNAUTHORIZED)
-			else:
-				update_last_connection(user)
-				return Response(UserSerializer(user).data, status=status.HTTP_200_OK)
-
-		except Usuario.DoesNotExist:
-			new_user_data = Usuario.objects.create(uid=user_uid, nombre=name, latitud_registro=latitud_registro, longitud_registro=longitud_registro)
-			return Response(UserSerializer(new_user_data).data, status=status.HTTP_201_CREATED)
-
+		else
+			return retorno
 
 @api_view(['GET'])
 @permission_classes((permissions.AllowAny,))
@@ -110,22 +77,10 @@ def GenericProductView(request, format=None):
 @permission_classes((permissions.AllowAny,))
 def GetUserProfile(request, format=None):
 	token = request.POST.get('token', 'nothing')
-	#auth.refresh(token)
 	user_uid = request.POST.get('uid', 'nothing')
-	if request.method != 'POST':
-		return Response(status=status.HTTP_400_BAD_REQUEST)
-	if token == 'nothing' or user_uid == 'nothing':
-		return Response(status=status.HTTP_400_BAD_REQUEST)
-	else:
-		#if check_user_logged_in(token):
-			#try:
-				fetch_user = Usuario.objects.get(uid=user_uid)
-				return Response(UserProfileSerializer(fetch_user).data, status=status.HTTP_200_OK)
-			#except:
-			#	return Response(status=status.HTTP_404_NOT_FOUND)
-		#else:
-		#	return Response(status=status.HTTP_401_UNAUTHORIZED)
-
+	if request.method != 'POST' or token == 'nothing' or user_uid == 'nothing': 
+			return Response(status=status.HTTP_400_BAD_REQUEST)
+	if retorno = 'Error':
 
 @api_view(['POST'])
 @permission_classes((permissions.AllowAny,))
